Fix Refactor Inheritance
