+++ backend/src/main/java/com/benine/backend/camera/FocussingCamera.java
@@ -4,7 +4,7 @@ package com.benine.backend.camera;
  * Decorator of a camera with functions to control the Focussing of the camera.
  * @author Bryan
  */
-public interface FocussingCamera extends ControlableCamera {
+public interface FocussingCamera extends ControllableCamera {
   
   /**
    * Get the focus position.
