+++ backend/src/main/java/com/benine/backend/camera/MovingCamera.java
@@ -4,7 +4,7 @@ package com.benine.backend.camera;
  * Decorator of a camera with functions to control the movements of the camera.
  * @author Bryan
  */
-public interface MovingCamera extends ControlableCamera {
+public interface MovingCamera extends ControllableCamera {
   
   /**
    * Move the camera to the specified position.
